Finally got the telemetry to work
