+++ firmware/HC12Tracker.ewp
@@ -1652,6 +1652,12 @@
         <name>$PROJ_DIR$\spi_bitbang.h</name>
     </file>
     <file>
+        <name>$PROJ_DIR$\string.c</name>
+    </file>
+    <file>
+        <name>$PROJ_DIR$\string.h</name>
+    </file>
+    <file>
         <name>$PROJ_DIR$\telemetry.c</name>
     </file>
     <file>
@@ -1659,5 +1665,8 @@
     </file>
     <file>
         <name>$PROJ_DIR$\tlm.c</name>
+    </file>
+    <file>
+        <name>$PROJ_DIR$\tlm.h</name>
     </file>
 </project>
