+++ firmware/tlm.h
@@ -1,0 +1,17 @@
+//#ifndef __tlm_h__
+//#define __tlm_h__
+//
+//#include <inttypes.h>
+//#include "main.h"
+//
+///* because tick is slower when using DominoEX,
+// * use less ticks for blipping
+// */
+//#define	BLIP_FACTOR	5
+//
+//void tx_blips(uint8_t reset);
+//void init_tx_buffer(void);
+//void prepare_tx_buffer(void);
+//void tlm_init(void);
+//
+//#endif
